docs: move validations to it's own file
